split github concerns into a real lib file system
