+++ lib/github.rb
@@ -1,0 +1,8 @@
+module GitHub
+  class Error < StandardError; end
+  class Forbidden < Error; end
+  class NotFound < Error; end
+
+  autoload :APIHeaders, 'github/api_headers'
+  autoload :Errors,     'github/errors'
+end
