+++ lib/github/api_headers.rb
@@ -1,0 +1,9 @@
+module GitHub
+  module APIHeaders
+    class << self
+      def no_cache_no_store
+        { 'Cache-Control' => 'no-cache, no-store' }
+      end
+    end
+  end
+end
